requirements.txt
fastapi
uvicorn[standard]
bcrypt
pymongo
PyJWT
python-decouple
pydantic[email]